2to3: long

+++ src/sardana/macroserver/recorders/sharedmemory.py
@@ -143,7 +143,7 @@ class SPSRecorder(BaseSharedMemoryRecorder):
 
         for colname in self.labels:
             val = record.data.get(colname)
-            if (not val is None) and (operator.isNumberType(val) and (type(val) in [int, float, long])):
+            if (not val is None) and (operator.isNumberType(val) and (type(val) in [int, float, int])):
                 vals.append(val)
             elif (not val is None) and (operator.isNumberType(val)):
                 valsmca = []
@@ -271,7 +271,7 @@ class ShmRecorder(DataRecorder):
         for colname in self.labels:
             dim_list.append(0)
             val = record.data.get(colname)
-            if (not val is None) and (type(val) in [int, float, long]):
+            if (not val is None) and (type(val) in [int, float, int]):
                 vals.append(val)
 
         myj = 0
